[BUG FIX] Duplicate Catalogs when updating checkin via summary page
When a user changes the value of a condition/symptom catalog it would
produce a duplicate on the next app launch.

As when a new question is appended to the fragment it checks if it is
already in the catalog definition…. Due to the refactor it didn’t think
that it was in the catalog definition and added anyway.

----

Also deletion of trackable would not update in summary. As the API does not delete information from the checkin when deleting a trackable a seperate API call is needed to update the checkin, removing the trackable for that date.