Displaying input views in check in, subscribed to observer, prevent duplicate items.
Input views are displayed inside the check in.

The questions are added/removed when the model changed. (Still need to
handle new input and updating the input when the value is changed).

Duplicate views where displayed when the fragment is recreated (back
tracing can cause this).

Bug when selecting an input value it crashes the app, this is because it needs changing to the new observer pattern.