Fixed bug where first check in pages could appear blank
It appears every other check in the first three or so pages are blank
(not the summary view)… I believe this could of been caused by
upgrading the android version to 23. Swiping through the pages and
returning to the start seems to make the check in appear again.
StackOverflow recommended changing the ViewPagerAdapter’s extending
class from FragmentPagerAdapter to FragmentStatePagerAdapter